% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_lm_predict_sims.R
\name{add_fitted_sims.lm}
\alias{add_fitted_sims.lm}
\title{Generate fitted simulations from a lm model}
\usage{
\method{add_fitted_sims}{lm}(newdata, mod, n_sims = 1000, seed = NULL, ...)
}
\arguments{
\item{newdata}{a data.frame of new data to predict}

\item{mod}{A \link{lm} model to simulate from.}

\item{n_sims}{number of simulation samples to construct}

\item{seed}{numeric, optional argument to set seed for simulations}

\item{...}{Unused dots for compatibility with generic functions.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about simulate values.
}
\description{
Generate simulations
from a lm model incorporating either error in
fitted error. Simulations  explore the possible space
of what a model might predict rather than an interval for use
in comparison to Bayesian posteriors for non-Bayesian models. The
output format and functions draw inspiration from the
\code{\link[tidybayes:tidybayes-package]{tidybayes::tidybayes()}} library and
\code{\link[merTools:predictInterval]{merTools::predictInterval()}}
}
\examples{
clotting <- data.frame(
   u = c(5,10,15,20,30,40,60,NA,100),
   lot1 = c(118,58,42,35,27,25,21,19,18),
   lot2 = c(69,35,26,21,18,16,13,12,12))

mod <- lm(lot1 ~ log(u) + lot2, data = clotting)

sims_fit <- add_fitted_sims(clotting, mod)

head(sims_fit)

}
\seealso{
Other lm: 
\code{\link{add_predicted_sims.lm}()}
}
\concept{lm}
