% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewisesem_psem_predict_sims.R
\name{add_fitted_sims.psem}
\alias{add_fitted_sims.psem}
\title{Generate fitted simulations from a psem model}
\usage{
\method{add_fitted_sims}{psem}(newdata, mod, n_sims = 1000, seed = NULL, ...)
}
\arguments{
\item{newdata}{a data.frame of new data to predict}

\item{mod}{An \code{\link[piecewiseSEM:psem]{piecewiseSEM::psem()}} object to simulate from.}

\item{n_sims}{number of simulation samples to construct}

\item{seed}{numeric, optional argument to set seed for simulations}

\item{...}{Unused dots for compatibility with generic functions.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about simulate values.
}
\description{
Generate simulations
from a psem model incorporating either error in
fitted error. Simulations  explore the possible space
of what a model might predict rather than an interval for use
in comparison to Bayesian posteriors for non-Bayesian models. The
output format and functions draw inspiration from the
\code{\link[tidybayes:tidybayes-package]{tidybayes::tidybayes()}} library and
\code{\link[merTools:predictInterval]{merTools::predictInterval()}}
}
\examples{

\dontrun{

library(piecewiseSEM)
data(keeley)
mod <- psem(
 lm(abiotic ~ distance, data=keeley),
 lm(rich ~ abiotic + hetero, data=keeley),
 lm(hetero ~ distance, data=keeley),
 data = keeley)

newdat <- data.frame(distance=c(30, 50))

new_fitted_sims <- newdat \%>\%
 add_fitted_sims(mod)

head(new_fitted_sims)
}
}
\seealso{
Other psem: 
\code{\link{add_predicted_sims.psem}()}
}
\concept{psem}
