% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sims.R
\name{add_fitted_sims}
\alias{add_fitted_sims}
\title{Generate fitted simulations from a fit model}
\usage{
add_fitted_sims(newdata, mod, n_sims = 1000, seed = NULL, ...)
}
\arguments{
\item{newdata}{a data.frame of new data to predict}

\item{mod}{An model to simulate from.}

\item{n_sims}{number of simulation samples to construct}

\item{seed}{numeric, optional argument to set seed for simulations}

\item{...}{Additional arguments to prediction simulation method.}
}
\description{
Generate simulations
from a fit model incorporating either error in
fitted error. Simulations  explore the possible space
of what a model might predict rather than an interval for use
in comparison to Bayesian posteriors for non-Bayesian models. The
output format and functions draw inspiration from the
\code{\link[tidybayes:tidybayes-package]{tidybayes::tidybayes()}} library and
\code{\link[merTools:predictInterval]{merTools::predictInterval()}}
}
