% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewisesem_psem_predict_sims.R
\name{add_predicted_sims.psem}
\alias{add_predicted_sims.psem}
\title{Generate predicted simulations from a psem model}
\usage{
\method{add_predicted_sims}{psem}(
  newdata,
  mod,
  n_sims = 1000,
  seed = NULL,
  type = c("predict", "fit"),
  ...
)
}
\arguments{
\item{newdata}{a data.frame of new data to predict}

\item{mod}{An \code{\link[piecewiseSEM:psem]{piecewiseSEM::psem()}} object to simulate from.}

\item{n_sims}{number of simulation samples to construct}

\item{seed}{numeric, optional argument to set seed for simulations}

\item{type}{Character defining if we are looking at fit or predict intervals.}

\item{...}{Unused dots for compatibility with generic functions.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about simulate values.
}
\description{
Generate simulations
from a psem model incorporating either error in
predicted error. Simulations  explore the possible space
of what a model might predict rather than an interval for use
in comparison to Bayesian posteriors for non-Bayesian models. The
output format and functions draw inspiration from the
\code{\link[tidybayes:tidybayes-package]{tidybayes::tidybayes()}} library and
\code{\link[merTools:predictInterval]{merTools::predictInterval()}}
}
\examples{

\dontrun{

library(piecewiseSEM)
data(keeley)
mod <- psem(
 lm(abiotic ~ distance, data=keeley),
 lm(rich ~ abiotic + hetero, data=keeley),
 lm(hetero ~ distance, data=keeley),
 data = keeley)

newdat <- data.frame(distance=c(30, 50))

new_predicted_sims <- newdat \%>\%
 add_predicted_sims(mod)

head(new_predicted_sims)
}
}
\seealso{
Other psem: 
\code{\link{add_fitted_sims.psem}()}
}
\concept{psem}
